pyproject.toml
[project]
name = "weather-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "load-dotenv>=0.1.0",
    "numpy>=2.4.4",
    "openmeteo-requests>=1.7.5",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "requests>=2.33.1",
    "requests-cache>=1.3.1",
    "retry>=0.9.2",
    "retry-requests>=2.0.0",
]
